Reusable addmapserver
